+++ frontend/src/constants.ts
@@ -4,6 +4,9 @@ export const constants = {
     URL_PREFIX: "http://localhost:7000/api",
     SCREEN_WIDTH: 960,
     SCREEN_HEIGHT: 720,
-    // Colours
+    // Colours (darkest to lightest gray at top, then ROYGBIV going downwards below that)
     APP_BACKGROUND: "#03090A",
+    EGGSHELL: "19191F",
+    RED_ERROR: "#D10000",
+    GREEN_TERMINAL: "#009A00",
 }